Update the ems-intune-govt-service-description.md
CSS got several cases from customer reporting the WCD failed to fetch the AAD bulk token. Engineering is working on a fix but that could take months since involves a change in AAD, WCD tool, and the OS platform to fetch the token and support this enrollment method.

Please add the example line (36)until the fix is developed: Bulk Enrollment using the Windows Configuration Designer is not currently supported in GCC High.